@@ -2,15 +2,13 @@
 
 namespace Jumbo\Client\Factory\Client;
 
-use ReflectionClass;
-
-use Zend\ServiceManager\AbstractFactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
-
 use Jumbo\Client\JumboClient;
 use Jumbo\Client\Exception;
 use Jumbo\Client\Options\Client\ClientOptions;
 use Jumbo\Client\Options\ModuleOptions;
+use Zend\ServiceManager\AbstractFactoryInterface;
+use Zend\ServiceManager\ServiceLocatorInterface;
+use ReflectionClass;
 
 class ClientFactory implements
     AbstractFactoryInterface
@@ -20,12 +18,11 @@ class ClientFactory implements
      *
      * @var array
      */
-    protected $lookupCache = array();
+    private $lookupCache = [];
 
     /**
      * Determine if we can create a service with name.
      *
-     * @param ServiceLocatorInterface $services
      * @param string $name
      * @param string $requestedName
      * @return boolean
@@ -46,7 +43,6 @@ class ClientFactory implements
     /**
      * Create service with name.
      *
-     * @param ServiceLocatorInterface $services
      * @param string $name
      * @param string $requestedName
      * @return JumboClient
@@ -56,14 +52,14 @@ class ClientFactory implements
         $clientOptions = $this->getClientOptions($services, $requestedName);
 
         if (!$this->clientExists($requestedName)) {
-            throw new Exception\ConfigException(
+            throw new Exception\RuntimeException(
                 sprintf('Client "%s" does not exist', $requestedName)
             );
         }
 
         /** @var JumboClient $requestedName*/
 
-        $appliedClientOptions = array();
+        $appliedClientOptions = [];
 
         // Only pass along options which are actually set
         if ($clientOptions !== null) {
@@ -77,11 +73,7 @@ class ClientFactory implements
         return $client;
     }
 
-    /**
-     * @param string $clientClass
-     * @return boolean
-     */
-    protected function clientExists($clientClass)
+    private function clientExists(string $clientClass): bool
     {
         // Class name must start with "Jumbo\Client"
         if (strpos($clientClass, 'Jumbo\Client') !== 0) {
@@ -97,12 +89,7 @@ class ClientFactory implements
         return $reflectionClass->isSubclassOf(JumboClient::CLASS);
     }
 
-    /**
-     * @param ServiceLocatorInterface $services
-     * @param string $clientName
-     * @return ClientOptions
-     */
-    protected function getClientOptions(ServiceLocatorInterface $services, $clientName)
+    private function getClientOptions(ServiceLocatorInterface $services, string $clientName): ClientOptions
     {
         /** @var ModuleOptions $moduleOptions */
         $moduleOptions = $services->get(ModuleOptions::CLASS);
