@@ -2,20 +2,14 @@
 
 namespace Jumbo\Client\Options;
 
-use Detail\Core\Options\AbstractOptions;
+use Zend\Stdlib\AbstractOptions;
 
 class ModuleOptions extends AbstractOptions
 {
-    /**
-     * @var Client\ClientOptions[]
-     */
-    protected $clients = array();
+    /** @var Client\ClientOptions[] */
+    private $clients = [];
 
-    /**
-     * @param string $name
-     * @return Client\ClientOptions
-     */
-    public function getClient($name)
+    public function getClient(string $name): ?Client\ClientOptions
     {
         return isset($this->clients[$name]) ? $this->clients[$name] : null;
     }
@@ -23,7 +17,7 @@ class ModuleOptions extends AbstractOptions
     /**
      * @return Client\ClientOptions[]
      */
-    public function getClients()
+    public function getClients(): array
     {
         return $this->clients;
     }
@@ -31,9 +25,9 @@ class ModuleOptions extends AbstractOptions
     /**
      * @param array $clients
      */
-    public function setClients(array $clients)
+    public function setClients(array $clients): void
     {
-        $this->clients = array();
+        $this->clients = [];
 
         foreach ($clients as $name => $options) {
             $this->clients[$name] = new Client\ClientOptions($options);
