@@ -7,36 +7,34 @@ use Jumbo\Client\Options\ModuleOptions;
 
 class ModuleOptionsTest extends OptionsTestCase
 {
-    /**
-     * @var ModuleOptions
-     */
+    /** @var ModuleOptions */
     protected $options;
 
     protected function setUp()
     {
         $this->options = $this->getOptions(
             ModuleOptions::CLASS,
-            array(
+            [
                 'getClient',
                 'getClients',
                 'setClients',
-            )
+            ]
         );
     }
 
-    public function testOptionsExist()
+    public function testOptionsExist(): void
     {
         $this->assertInstanceOf(ModuleOptions::CLASS, $this->options);
     }
 
-    public function testClientCanBeSet()
+    public function testClientCanBeSet(): void
     {
         $clientName = 'SomeClient';
-        $clientsConfig = array(
-            $clientName => array(
-                'base_url' => 'http://some.client.com',
-            ),
-        );
+        $clientsConfig = [
+            $clientName => [
+                'base_uri' => 'http://some.client.com',
+            ],
+        ];
 
         $this->assertTrue(is_array($this->options->getClients()));
         $this->assertCount(0, $this->options->getClients());
