@@ -6,23 +6,18 @@ use Jumbo\Client\AssetsClient as Client;
 
 trait AssetsClientAwareTrait
 {
-    /**
-     * @var Client
-     */
+    /** @var Client */
     protected $assetsClient;
 
     /**
      * @return Client
      */
-    public function getAssetsClient()
+    public function getAssetsClient(): Client
     {
         return $this->assetsClient;
     }
 
-    /**
-     * @param Client $assetsClient
-     */
-    public function setAssetsClient(Client $assetsClient)
+    public function setAssetsClient(Client $assetsClient): void
     {
         $this->assetsClient = $assetsClient;
     }
